% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_phenotype_analysis.R
\name{all_module_TTESTPAIRED_test}
\alias{all_module_TTESTPAIRED_test}
\title{Title all_module_TTESTPAIRED_test}
\usage{
all_module_TTESTPAIRED_test(
  AMARETTOinit,
  AMARETTOresults,
  test_sample_annotations
)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}

\item{test_sample_annotations}{}
}
\value{

}
\description{
Title all_module_TTESTPAIRED_test
}
