% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_phenotype_analysis.R
\name{all_modules_PEARSONCOR_test}
\alias{all_modules_PEARSONCOR_test}
\title{Title}
\usage{
all_modules_PEARSONCOR_test(AMARETTOinit, AMARETTOresults,
  test_sample_annotations)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}

\item{test_sample_annotations}{}
}
\description{
Title
}
