% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_run.R
\name{AMARETTO_CreateRegulatorPrograms}
\alias{AMARETTO_CreateRegulatorPrograms}
\title{AMARETTO_CreateRegulatorPrograms}
\usage{
AMARETTO_CreateRegulatorPrograms(AMARETTOinit, AMARETTOresults)
}
\arguments{
\item{AMARETTOinit}{List output from AMARETTO_Initialize().}

\item{AMARETTOresults}{List output from AMARETTO_Run()}
}
\value{
result
}
\description{
AMARETTO_CreateRegulatorPrograms
}
\examples{
data('ProcessedDataLIHC')
AMARETTOinit <- AMARETTO_Initialize(ProcessedData = ProcessedDataLIHC,
                                    NrModules = 2, VarPercentage = 50)
AMARETTOresults <- AMARETTO_Run(AMARETTOinit)
AMARETTO_RP <- AMARETTO_CreateRegulatorPrograms(AMARETTOinit,AMARETTOresults)
}
