% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_download.R
\name{AMARETTO_Download}
\alias{AMARETTO_Download}
\title{AMARETTO_Download}
\usage{
AMARETTO_Download(CancerSite = "CHOL",
  TargetDirectory = TargetDirectory)
}
\arguments{
\item{CancerSite}{TCGA cancer code for data download}

\item{TargetDirectory}{Directory path to download data}
}
\value{
result
}
\description{
Downloading TCGA dataset for AMARETTO analysis
}
\examples{
TargetDirectory <- file.path(getwd(),"Downloads/");dir.create(TargetDirectory)
CancerSite <- 'CHOL'
DataSetDirectories <- AMARETTO_Download(CancerSite,TargetDirectory = TargetDirectory)
}
