% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_download.R
\name{AMARETTO_DownloadResults}
\alias{AMARETTO_DownloadResults}
\title{AMARETTO_DownloadResults
Retrieve a download of all the data linked with the run (including heatmaps)}
\usage{
AMARETTO_DownloadResults(AMARETTOinit, AMARETTOresults, data_address,
  Heatmaps = TRUE)
}
\arguments{
\item{AMARETTOinit}{AMARETTO initialize output}

\item{AMARETTOresults}{AMARETTO results output}

\item{data_address}{Directory to save data folder}

\item{Heatmaps}{Output heatmaps as pdf}
}
\description{
AMARETTO_DownloadResults
Retrieve a download of all the data linked with the run (including heatmaps)
}
\examples{
AMARETTO_DownloadResults(AMARETTOinit,AMARETTOresults,"./")
}
