% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_html.R
\name{amaretto_html_report}
\alias{amaretto_html_report}
\title{Title}
\usage{
amaretto_html_report(AMARETTOinit, AMARETTOresults, CNV_matrix, MET_matrix,
  VarPercentage, hyper_geo_test_bool = TRUE,
  n_cluster = AMARETTOinit$NrCores, wordcloud_bool = FALSE,
  hyper_geo_refence_name = "H.C2CP.genesets.gmt",
  output_address = "./", output_name = "cancer")
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}

\item{CNV_matrix}{}

\item{MET_matrix}{}

\item{VarPercentage}{}

\item{hyper_geo_test_bool}{}

\item{n_cluster}{}

\item{wordcloud_bool}{}

\item{hyper_geo_refence_name}{}

\item{output_address}{}

\item{output_name}{}
}
\description{
Title
}
