% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlphenotypes.R
\name{AMARETTO_HTMLCharacterisation}
\alias{AMARETTO_HTMLCharacterisation}
\title{AMARETTO_HTMLCharacterisation}
\usage{
AMARETTO_HTMLCharacterisation(AMARETTOresults, annotation,
  report_address = "./", survivalanalysis = FALSE,
  phenotypeanalysis = FALSE, idpatients = NULL, idvitalstatus = NULL,
  iddaystodeath = NULL, idfollowupdays = NULL, daystoyears = FALSE,
  rightcensored = FALSE, parameters = NULL, typelist = NULL)
}
\arguments{
\item{printsurv}{}
}
\value{
A set of HTMLs, giving caracteristics of the communities
}
\description{
AMARETTO_HTMLCharacterisation
}
