% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{AMARETTO_HTMLreport}
\alias{AMARETTO_HTMLreport}
\title{AMARETTO_HTMLreport}
\usage{
AMARETTO_HTMLreport(AMARETTOinit, AMARETTOresults, CNV_matrix = NULL,
  MET_matrix = NULL, SAMPLE_annotation = NULL, ID = NULL,
  VarPercentage, hyper_geo_test_bool = FALSE,
  hyper_geo_reference = NULL, output_address = "./", MSIGDB = FALSE,
  GMTURL = FALSE)
}
\arguments{
\item{AMARETTOinit}{AMARETTO initialize output}

\item{AMARETTOresults}{AMARETTO results output}

\item{CNV_matrix}{Processed CNV matrix ($CNV_matrix element from Preprocess_CancerSite() list output).}

\item{MET_matrix}{Processed methylation matrix ($MET_matrix element from Preprocess_CancerSite() list output).}

\item{SAMPLE_annotation}{SAMPLE annotation will be added to heatmap}

\item{ID}{ID column of the SAMPLE annotation data frame}

\item{VarPercentage}{Original Var Percentage used}

\item{hyper_geo_test_bool}{Boolean if a hyper geometric test needs to be performed. If TRUE provide a GMT file in the hyper_geo_reference parameter.}

\item{hyper_geo_reference}{GMT file with gene sets to compare with.}

\item{output_address}{Output directory for the html files.}

\item{MSIGDB}{TRUE if gene sets were retrieved from MSIGDB. Links will be created in the report.}

\item{GMTURL}{TRUE if second column of gmt contains URLs to gene set, FALSE if it contains a description}
}
\description{
Retrieve an interactive html report, including gene set enrichment analysis if asked for.
}
\examples{
data('ProcessedDataLIHC')
AMARETTOinit <- AMARETTO_Initialize(MA_matrix = ProcessedDataLIHC$MA_matrix,
                                    CNV_matrix = ProcessedDataLIHC$CNV_matrix,
                                    MET_matrix = ProcessedDataLIHC$MET_matrix,
                                    NrModules = 20, VarPercentage = 50)

AMARETTOresults <- AMARETTO_Run(AMARETTOinit)

AMARETTO_HTMLreport(AMARETTOinit= AMARETTOinit,AMARETTOresults= AMARETTOresults,
                    CNV_matrix=ProcessedDataLIHC$CNV_matrix,
                    MET_matrix = ProcessedDataLIHC$MET_matrix,
                    VarPercentage=10,hyper_geo_test_bool=FALSE,
                    output_address='./')

}
