% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_run.R
\name{AMARETTO_LarsenBased}
\alias{AMARETTO_LarsenBased}
\title{AMARETTO_LarsenBased}
\usage{
AMARETTO_LarsenBased(Data, Clusters, RegulatorData, Parameters, NrCores)
}
\arguments{
\item{Data}{}

\item{Clusters}{}

\item{RegulatorData}{}

\item{Parameters}{}

\item{NrCores}{}
}
\description{
AMARETTO_LarsenBased
}
