% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_run.R
\name{AMARETTO_LearnRegulatoryPrograms}
\alias{AMARETTO_LearnRegulatoryPrograms}
\title{AMARETTO_LearnRegulatoryPrograms}
\usage{
AMARETTO_LearnRegulatoryPrograms(Data, Clusters, RegulatorData,
  RegulatorSign, Lambda, AutoRegulation, Parameters)
}
\arguments{
\item{mode}{}
}
\description{
AMARETTO_LearnRegulatoryPrograms
}
\keyword{internal}
