% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_phenotypes.R
\name{AMARETTO_PhenAssociation}
\alias{AMARETTO_PhenAssociation}
\title{AMARETTO_PhenoAssociation}
\usage{
AMARETTO_PhenAssociation(AMARETTOresults, annotation, idpatients,
  phenotypetests, printplots = FALSE, pdfname = "phenotypes.pdf")
}
\arguments{
\item{pdfname}{}
}
\description{
AMARETTO_PhenoAssociation
}
