% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_phenotypes.R
\name{AMARETTO_phenotypesshiny}
\alias{AMARETTO_phenotypesshiny}
\title{AMARETTO_phenotypesshiny}
\usage{
AMARETTO_phenotypesshiny(AMARETTOresults, annotation, idpatients)
}
\arguments{
\item{annotation}{}
}
\value{
A set of HTMLs, giving caracteristics of the communities


}
\description{
AMARETTO_phenotypesshiny
}
