% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{AMARETTO_Preprocess}
\alias{AMARETTO_Preprocess}
\title{AMARETTO_Preprocess}
\usage{
AMARETTO_Preprocess(DataSetDirectories = DataSetDirectories,
  BatchData = BatchData)
}
\arguments{
\item{DataSetDirectories}{DataSetDirectories}

\item{BatchData}{BatchData}
}
\value{
result
}
\description{
Wrapper code that analyzes process TCGA GISTIC (CNV) and gene expression (rna-seq or microarray) data  via one call
}
\examples{
\dontrun{
data('BatchData')
TargetDirectory <-  tempdir()  # path to data download directory
CancerSite <- 'CHOL'
DataSetDirectories <- AMARETTO_Download(CancerSite,TargetDirectory)
ProcessedData <- AMARETTO_Preprocess(DataSetDirectories)
}
}
