% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_run.R
\name{AMARETTO_ReassignGenesToClusters}
\alias{AMARETTO_ReassignGenesToClusters}
\title{AMARETTO_ReassignGenesToClusters}
\usage{
AMARETTO_ReassignGenesToClusters(
  Data,
  RegulatorData,
  Beta,
  Clusters,
  AutoRegulation
)
}
\value{
result
}
\description{
AMARETTO_ReassignGenesToClusters
}
\keyword{internal}
