% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_phenotype_analysis.R
\name{AMARETTO_statistical_test}
\alias{AMARETTO_statistical_test}
\title{Title}
\usage{
AMARETTO_statistical_test(AMARETTOinit, AMARETTOresults,
  sample_annotations, phenotype_tests)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}

\item{sample_annotations}{}

\item{phenotype_tests}{}
}
\description{
Title
}
