% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_survival.R
\name{AMARETTO_survival}
\alias{AMARETTO_survival}
\title{AMARETTO_survival}
\usage{
AMARETTO_survival(AMARETTOresults, SurvivalAnnotation, idpatients,
  idvitalstatus, iddaystodeath, idfollowupdays, daystoyears,
  rightcensored = FALSE, printHR = FALSE, printsurv = FALSE)
}
\arguments{
\item{printsurv}{}
}
\value{
A set of HTMLs, giving caracteristics of the communities
}
\description{
AMARETTO_survival
}
