% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_survival.R
\name{AMARETTO_survivalshiny}
\alias{AMARETTO_survivalshiny}
\title{AMARETTO_survivalshiny}
\usage{
AMARETTO_survivalshiny(AMARETTOresults, SurvivalAnnotation, idpatients,
  idvitalstatus, iddaystodeath, idfollowupdays)
}
\arguments{
\item{rightcensored}{}
}
\value{
A set of HTMLs, giving caracteristics of the communities
}
\description{
AMARETTO_survivalshiny
}
