% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\docType{class}
\name{AmReport-class}
\alias{AmReport-class}
\title{AMARETTO report data structure}
\description{
AMARETTO report data structure
}
\section{Slots}{

\describe{
\item{\code{nExp}}{}

\item{\code{nCNV}}{}

\item{\code{nMET}}{}

\item{\code{nGenes}}{}

\item{\code{VarPercentage}}{}

\item{\code{nMod}}{}
}}

