% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{AmReportFunction}
\alias{AmReportFunction}
\title{constructor for AmReport instances, built from list}
\usage{
AmReportFunction(reportlist)
}
\arguments{
\item{reportlist}{a list with obligatory elements}
}
\description{
constructor for AmReport instances, built from list
}
\examples{
repobj = AmReport(AmObjects::LIHC_AMARETTOreport)
repobj
}
