% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{AmReportModuleFunction}
\alias{AmReportModuleFunction}
\title{constructor for AmReport instances, built from list}
\usage{
AmReportModuleFunction(reportlist_module)
}
\arguments{
\item{reportlist_module}{}
}
\description{
constructor for AmReport instances, built from list
}
\examples{
repobj = AmReport(AmObjects::LIHC_AMARETTOreport)
repobj
}
