% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BatchData}
\alias{BatchData}
\title{BatchData}
\format{An object of class \code{data.frame} with 23263 rows and 3 columns.}
\source{
\url{AMARETTO}
}
\usage{
BatchData
}
\description{
A dataset for conducting batch corerction in TCGA samples
}
\keyword{datasets}
