% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{build.design}
\alias{build.design}
\title{build.design}
\usage{
build.design(vec, des = NULL, start = 2)
}
\value{
result
}
\description{
Next two functions make the design matrix (X) from the sample info file
}
\keyword{internal}
