% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_download.R
\name{cacheResource}
\alias{cacheResource}
\title{cacheResource}
\usage{
cacheResource(TargetDirectory = TargetDirectory, resource = resource)
}
\value{
result
}
\description{
cacheResource
}
\keyword{internal}
