% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_download.R
\name{computeGisticURL}
\alias{computeGisticURL}
\title{computeGisticURL}
\usage{
computeGisticURL(url = NULL, acronym = "CHOL")
}
\arguments{
\item{url}{}

\item{acronym}{}
}
\value{
result
}
\description{
computeGisticURL
}
\keyword{internal}
