% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_phenotype_analysis.R
\name{cox_prop_hazard_test}
\alias{cox_prop_hazard_test}
\title{Title}
\usage{
cox_prop_hazard_test(xx, time_to_event, right_censoring)
}
\arguments{
\item{xx}{}

\item{time_to_event}{}

\item{right_censoring}{}
}
\description{
Title
}
