% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_phenotype_analysis.R
\name{create_feature_matrix}
\alias{create_feature_matrix}
\title{Title create_feature_matrix}
\usage{
create_feature_matrix(
  AMARETTOinit,
  AMARETTOresults,
  sample_annotation_df,
  module_number,
  Sample = "Sample"
)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}

\item{sample_annotation_df}{}

\item{module_number}{}

\item{Sample}{}
}
\value{

}
\description{
Title create_feature_matrix
}
