% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_html.R
\name{create_gene_annotations}
\alias{create_gene_annotations}
\title{Title}
\usage{
create_gene_annotations(module_all_genes_data, Module_Genes_names,
  Module_regulators_weights)
}
\arguments{
\item{module_all_genes_data}{}

\item{Module_Genes_names}{}

\item{Module_regulators_weights}{}
}
\description{
Title
}
