% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R,
%   R/amaretto_htmlreport_imaging.R
\name{create_hgt_datatable}
\alias{create_hgt_datatable}
\title{Title create_hgt_datatable}
\usage{
create_hgt_datatable(output_hgt, module_table, ModuleNr = 1)

create_hgt_datatable(output_hgt, module_table, ModuleNr = 1)
}
\arguments{
\item{output_hgt}{GSEA test dataframe from HyperGeoEnrichmentTest function.}

\item{module_table}{If TRUE, makes the ModuleNr datatable, If FALSE, makes the index page datatable.}

\item{ModuleNr}{The module number.}
}
\value{
result

result
}
\description{
Title create_hgt_datatable

Title create_hgt_datatable
}
