% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Driver_Genes}
\alias{Driver_Genes}
\title{Driver_Genes}
\format{List}
\source{
\url{AMARETTO}
}
\usage{
Driver_Genes
}
\description{
A list of cancer driver genes described in literature.
}
\keyword{datasets}
