% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{driver_genes_summary}
\alias{driver_genes_summary}
\title{Title driver_genes_summary
Provide summary of all driver genes in that}
\usage{
driver_genes_summary(AMARETTOresults, weight_threshold = 0.001,
  plot_wordcloud = FALSE)
}
\arguments{
\item{AMARETTOresults}{}

\item{weight_threshold}{}

\item{plot_wordcloud}{}
}
\value{
all_regulators_grouped
}
\description{
Title driver_genes_summary
Provide summary of all driver genes in that
}
