% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{filter.absent}
\alias{filter.absent}
\title{filter.absent}
\usage{
\method{filter}{absent}(x, pct)
}
\value{
result
}
\value{
result
}
\description{
filters data based on presence/absence call
}
\keyword{internal}
