% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_initialize.R
\name{FindTranscriptionallyPredictive_CNV}
\alias{FindTranscriptionallyPredictive_CNV}
\title{FindTranscriptionallyPredictive_CNV}
\usage{
FindTranscriptionallyPredictive_CNV(MA_matrix, CNV_matrix,
  PvalueThreshold = 0.001, RsquareThreshold = 0.1)
}
\value{
result
}
\description{
Function to identify which genes CNV significantly predict expression of that gene.
}
\keyword{internal}
