% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_initialize.R
\name{geneFiltering}
\alias{geneFiltering}
\title{geneFiltering
Function to filter gene expression matrix}
\usage{
geneFiltering(Type, MAdata, Percentage)
}
\arguments{
\item{Type}{}

\item{MAdata}{Processed gene expression data matrix ($MA_matrix element from Preprocess_CancerSite() list output).}

\item{Percentage}{Minimum percentage by variance for filtering of genes; for example, 75\% would indicate that the CreateRegulatorData() function only analyses genes that have a variance above the 75th percentile across all samples.}
}
\description{
geneFiltering
Function to filter gene expression matrix
}
