% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{genes_amaretto_assignments_all}
\alias{genes_amaretto_assignments_all}
\title{Title genes_amaretto_assignments_all}
\usage{
genes_amaretto_assignments_all(AMARETTOinit, AMARETTOresults)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}
}
\value{

}
\description{
Title genes_amaretto_assignments_all
}
