% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{GeneSetDescription}
\alias{GeneSetDescription}
\title{GeneSetDescription}
\usage{
GeneSetDescription(filename, MSIGDB)
}
\arguments{
\item{filename}{The name of the gmt file.}

\item{MSIGDB}{If True, the gene set description column will be provided from MSIGDB.}
}
\value{
result
}
\description{
GeneSetDescription
}
\keyword{internal}
