% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{GmtFromModules}
\alias{GmtFromModules}
\title{GmtFromModules}
\usage{
GmtFromModules(AMARETTOresults, driverGSEA)
}
\arguments{
\item{AMARETTOresults}{List output from AMARETTO_Run().}

\item{driverGSEA}{if TRUE , module driver genes will also be added to module target genes for GSEA.}
}
\value{
result
}
\description{
GmtFromModules
}
\keyword{internal}
