% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{HubAuthority_scores_AMARETTO}
\alias{HubAuthority_scores_AMARETTO}
\title{Title HubAuthority_scores_AMARETTO}
\usage{
HubAuthority_scores_AMARETTO(AMARETTOinit, AMARETTOresults)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}
}
\value{
a list containing Hub-score authority-score of genes and modules.
}
\description{
Title HubAuthority_scores_AMARETTO
}
\examples{
try(HubAuthority_scores_AMARETTO(AMARETTOinit,AMARETTOresults))
}
