% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{HyperGTestGeneEnrichment}
\alias{HyperGTestGeneEnrichment}
\title{Hyper Geometric Geneset Enrichement Test}
\usage{
HyperGTestGeneEnrichment(gmtfile, testgmtfile, NrCores,
  ref.numb.genes = 45956)
}
\arguments{
\item{gmtfile}{gmt with reference gene set}

\item{testgmtfile}{gmt with gene sets to test}

\item{NrCores}{number of cores for parallelization}

\item{ref.numb.genes}{The total number of genes teste, standard equal to 45 956 (MSIGDB standard)}
}
\description{
Hyper Geometric Geneset Enrichement Test
}
\keyword{internal}
