% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_html.R
\name{HyperGTestGeneEnrichment_serial}
\alias{HyperGTestGeneEnrichment_serial}
\title{Title}
\usage{
HyperGTestGeneEnrichment_serial(gmtfile, testgmtfile, outfile,
  show.overlapping.genes = FALSE, filter.genes = TRUE,
  show.unrecognized = FALSE)
}
\arguments{
\item{gmtfile}{}

\item{testgmtfile}{}

\item{outfile}{}

\item{show.overlapping.genes}{}

\item{filter.genes}{}

\item{show.unrecognized}{}
}
\description{
Title
}
