% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{int.eprior}
\alias{int.eprior}
\title{int.eprior}
\usage{
int.eprior(sdat, g.hat, d.hat)
}
\value{
result
}
\value{
result
}
\description{
Monte Carlo integration function to find the nonparametric adjustments
}
\keyword{internal}
