% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{it.sol}
\alias{it.sol}
\title{Title}
\usage{
it.sol(sdat, g.hat, d.hat, g.bar, t2, a, b, conv = 1e-04)
}
\arguments{
\item{sdat}{}

\item{g.hat}{}

\item{d.hat}{}

\item{g.bar}{}

\item{t2}{}

\item{a}{}

\item{b}{}

\item{conv}{}
}
\description{
Title
}
