% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{list.batch}
\alias{list.batch}
\title{list.batch}
\usage{
list.batch(saminfo)
}
\value{
result
}
\value{
result
}
\description{
Makes a list with elements pointing to which array belongs to which batch
}
\keyword{internal}
