% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MsigdbMapping}
\alias{MsigdbMapping}
\title{MsigdbMapping}
\format{An object of class \code{data.frame} with 17810 rows and 3 columns.}
\source{
\url{AMARETTO}
}
\usage{
MsigdbMapping
}
\description{
A dataframe containing of MSIGDB geneset names, their urls and descriptions.
}
\keyword{datasets}
