% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_phenotype_analysis.R
\name{phenotye_add_estimate_column}
\alias{phenotye_add_estimate_column}
\title{Title phenotye_add_estimate_column}
\usage{
phenotye_add_estimate_column(pheno_tbl)
}
\arguments{
\item{pheno_tbl}{}
}
\value{

}
\description{
Title phenotye_add_estimate_column
}
