% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{postmean}
\alias{postmean}
\title{Title}
\usage{
postmean(g.hat, g.bar, n, d.star, t2)
}
\arguments{
\item{g.hat}{}

\item{g.bar}{}

\item{n}{}

\item{d.star}{}

\item{t2}{}
}
\description{
Title
}
