% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_initialize.R
\name{printf}
\alias{printf}
\title{printf
Wrapper function for C-style formatted output.}
\usage{
printf(...)
}
\arguments{
\item{...}{}
}
\description{
printf
Wrapper function for C-style formatted output.
}
