% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_html.R
\name{processTCGA_modules}
\alias{processTCGA_modules}
\title{Title}
\usage{
processTCGA_modules(AMARETTOinit, AMARETTOresults)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}
}
\description{
Title
}
