% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_vizualize.R
\name{randomColor}
\alias{randomColor}
\title{randomColor}
\usage{
randomColor(count = 1, hue = c(" ", "random", "red", "orange",
  "yellow", "green", "blue", "purple", "pink", "monochrome"),
  luminosity = c(" ", "random", "light", "bright", "dark"))
}
\arguments{
\item{count}{count}

\item{hue}{hue}

\item{luminosity}{luminosity}
}
\value{
result
}
\description{
randomColor
}
\keyword{internal}
