% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R,
%   R/amaretto_htmlreport_imaging.R
\name{readGMT}
\alias{readGMT}
\title{readGMT}
\usage{
readGMT(filename)

readGMT(filename)
}
\arguments{
\item{filename}{}

\item{filename}{}
}
\value{
result

result
}
\description{
readGMT

readGMT
}
\keyword{internal}
