% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{Save_CancerSite}
\alias{Save_CancerSite}
\title{Title}
\usage{
Save_CancerSite(CancerSite, TargetDirectory, DataSetDirectories,
  ProcessedData)
}
\arguments{
\item{CancerSite}{}

\item{TargetDirectory}{}

\item{DataSetDirectories}{}

\item{ProcessedData}{}
}
\description{
Title
}
