% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_BatchCorrection_MolecularData}
\alias{TCGA_BatchCorrection_MolecularData}
\title{Title}
\usage{
TCGA_BatchCorrection_MolecularData(GEN_Data, BatchData, MinInBatch)
}
\arguments{
\item{GEN_Data}{}

\item{BatchData}{}

\item{MinInBatch}{}
}
\description{
Title
}
