% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_GENERIC_CheckBatchEffect}
\alias{TCGA_GENERIC_CheckBatchEffect}
\title{Title}
\usage{
TCGA_GENERIC_CheckBatchEffect(GEN_Data, BatchData)
}
\arguments{
\item{GEN_Data}{}

\item{BatchData}{}
}
\description{
Title
}
