% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_GENERIC_GetSampleGroups}
\alias{TCGA_GENERIC_GetSampleGroups}
\title{TCGA_GENERIC_GetSampleGroups}
\usage{
TCGA_GENERIC_GetSampleGroups(SampleNames)
}
\arguments{
\item{SampleNames}{}
}
\description{
TCGA_GENERIC_GetSampleGroups
}
\keyword{internal}
