% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_GENERIC_MergeData}
\alias{TCGA_GENERIC_MergeData}
\title{Title}
\usage{
TCGA_GENERIC_MergeData(NewIDListUnique, DataMatrix, MergeMethod)
}
\arguments{
\item{NewIDListUnique}{}

\item{DataMatrix}{}

\item{MergeMethod}{}
}
\description{
Title
}
