% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_Load_GISTICdata}
\alias{TCGA_Load_GISTICdata}
\title{TCGA_Load_GISTICdata}
\usage{
TCGA_Load_GISTICdata(GisticDirectory)
}
\arguments{
\item{GisticDirectory}{}
}
\description{
TCGA_Load_GISTICdata
}
\keyword{internal}
