% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_Load_MolecularData}
\alias{TCGA_Load_MolecularData}
\title{TCGA_Load_MolecularData}
\usage{
TCGA_Load_MolecularData(MAEO_ge)
}
\value{
result
}
\description{
TCGA_Load_MolecularData
}
\keyword{internal}
