% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{trim.dat}
\alias{trim.dat}
\title{trim.dat}
\usage{
trim.dat(dat)
}
\arguments{
\item{dat}{}
}
\value{
result
}
\description{
Trims the data of extra columns, note your array names cannot be named 'X' or start with 'X.'
}
\keyword{internal}
