% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_download.R
\name{write_gct}
\alias{write_gct}
\title{write_gct
Write a gct file with the proper header.}
\usage{
write_gct(data_in, file_address)
}
\arguments{
\item{file_address}{}
}
\description{
write_gct
Write a gct file with the proper header.
}
\examples{
write_gct(AMARETTOresults$RegulatoryPrograms,'RegulatoryPrograms.gct')
}
\keyword{internal}
