% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_run.R
\name{AMARETTO_CreateModuleData}
\alias{AMARETTO_CreateModuleData}
\title{AMARETTO_CreateModuleData}
\usage{
AMARETTO_CreateModuleData(AMARETTOinit, AMARETTOresults)
}
\arguments{
\item{AMARETTOresults}{}
}
\description{
AMARETTO_CreateModuleData
}
\keyword{internal}
