% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_run.R
\name{AMARETTO_CreateModuleData}
\alias{AMARETTO_CreateModuleData}
\title{Title}
\usage{
AMARETTO_CreateModuleData(AMARETTOinit, AMARETTOresults)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}
}
\description{
Title
}
