% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_run.R
\name{AMARETTO_CreateRegulatorPrograms}
\alias{AMARETTO_CreateRegulatorPrograms}
\title{AMARETTO_CreateRegulatorPrograms}
\usage{
AMARETTO_CreateRegulatorPrograms(AMARETTOinit, AMARETTOresults)
}
\arguments{
\item{AMARETTOresults}{}
}
\description{
AMARETTO_CreateRegulatorPrograms
}
\keyword{internal}
