% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_run.R
\name{AMARETTO_CreateRegulatorPrograms}
\alias{AMARETTO_CreateRegulatorPrograms}
\title{Title}
\usage{
AMARETTO_CreateRegulatorPrograms(AMARETTOinit, AMARETTOresults)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOresults}{}
}
\description{
Title
}
