% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_download.R
\name{AMARETTO_Download}
\alias{AMARETTO_Download}
\title{AMARETTO_Download}
\usage{
AMARETTO_Download(CancerSite, TargetDirectory, downloadData = TRUE)
}
\arguments{
\item{CancerSite}{TCGA cancer code for data download}

\item{TargetDirectory}{Directory path to download data}

\item{downloadData}{TRUE}
}
\value{
result
}
\description{
Downloading TCGA dataset for AMARETTO analysis
}
\examples{
CancerSite <- 'CHOL'
DataSetDirectories <- AMARETTO_Download(CancerSite,"./", FALSE)
}
