% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_download.R
\name{AMARETTO_DownloadResults}
\alias{AMARETTO_DownloadResults}
\title{AMARETTO_DownloadResults}
\usage{
AMARETTO_DownloadResults(AMARETTOinit, AMARETTOresults, data_address,
  Heatmaps = TRUE)
}
\arguments{
\item{AMARETTOinit}{AMARETTO initialize output}

\item{AMARETTOresults}{AMARETTO results output}

\item{data_address}{Directory to save data folder}

\item{Heatmaps}{Output heatmaps as pdf}

\item{data_in}{A dataframe that needs to be converted to a gct file}

\item{file_address}{File adress of the gct file}
}
\description{
Retrieve a download of all the data linked with the run (including heatmaps)
}
