% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_functions.R
\name{AMARETTO_EvaluateTestSet}
\alias{AMARETTO_EvaluateTestSet}
\title{AMARETTO_EvaluateTestSet}
\usage{
AMARETTO_EvaluateTestSet(AMARETTOresults, MA_Data_TestSet,
  RegulatorData_TestSet)
}
\arguments{
\item{AMARETTOresults}{AMARETTO output from AMARETTO_Run().}

\item{MA_Data_TestSet}{Gene expression matrix from a test set (that was not used in AMARETTO_Run()).}

\item{RegulatorData_TestSet}{Test regulator data from CreateRegulatorData().}
}
\description{
Code to evaluate AMARETTO on a new gene expression test set. Uses output from AMARETTO_Run() and CreateRegulatorData().
}
\examples{
AMARETTOinit <- AMARETTO_Initialize(MA_matrix=MA_matrix, CNV_matrix= CNV_matrix,MET_matrix= MET_matrix, Driver_list = NULL, NrModules= Nr, VarPercentage= Var,
                                    PvalueThreshold = 0.001, RsquareThreshold = 0.1, pmax = 10,
                                    NrCores = 1, OneRunStop = 0)

AMARETTOresults<-AMARETTO_Run(AMARETTOinit)
AMARETTOtestReport <- AMARETTO_EvaluateTestSet(AMARETTOresults,AMARETTOinit$MA_matrix_Var,AMARETTOinit$RegulatorData)
}
