% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{AMARETTO_htmlreport}
\alias{AMARETTO_htmlreport}
\title{AMARETTO_htmlreport}
\usage{
AMARETTO_htmlreport(AMARETTOinit, AMARETTOresults, CNV_matrix = NULL,
  MET_matrix = NULL, SAMPLE_annotation = NULL, ID = NULL,
  VarPercentage, hyper_geo_test_bool = FALSE,
  hyper_geo_reference = NULL, output_address = "./", MSIGDB = FALSE,
  GMTURL = FALSE)
}
\arguments{
\item{AMARETTOinit}{AMARETTO initialize output}

\item{AMARETTOresults}{AMARETTO results output}

\item{CNV_matrix}{Processed CNV matrix ($CNV_matrix element from Preprocess_CancerSite() list output).}

\item{MET_matrix}{Processed methylation matrix ($MET_matrix element from Preprocess_CancerSite() list output).}

\item{SAMPLE_annotation}{SAMPLE annotation will be added to heatmap}

\item{ID}{ID column of the SAMPLE annotation data frame}

\item{VarPercentage}{Original Var Percentage used}

\item{hyper_geo_test_bool}{Boolean if a hyper geometric test needs to be performed. If TRUE provide a GMT file in the hyper_geo_reference parameter.}

\item{hyper_geo_reference}{GMT file with gene sets to compare with.}

\item{output_address}{Output directory for the html files.}

\item{MSIGDB}{TRUE if gene sets were retrieved from MSIGDB. Links will be created in the report.}

\item{GMTURL}{TRUE if second column of gmt contains URLs to gene set, FALSE if it contains a description}
}
\description{
Retrieve an interactive html report, including gene set enrichment analysis if asked for.
}
\examples{
AMARETTO_htmlreport(AMARETTOinit,AMARETTOresults,CNV_matrix=CNV_matrix_LIHC,MET_matrix = MET_matrix_LIHC,VarPercentage=10,hyper_geo_test_bool=FALSE,output_address='./')
}
