% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_run.R
\name{AMARETTO_LearnRegulatoryProgramsLarsen}
\alias{AMARETTO_LearnRegulatoryProgramsLarsen}
\title{AMARETTO_LearnRegulatoryProgramsLarsen}
\usage{
AMARETTO_LearnRegulatoryProgramsLarsen(Data, Clusters, RegulatorData,
  RegulatorSign, Lambda, AutoRegulation, alpha, pmax)
}
\value{
result
}
\description{
AMARETTO_LearnRegulatoryProgramsLarsen
}
\keyword{internal}
