% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{AMARETTO_Preprocess}
\alias{AMARETTO_Preprocess}
\title{AMARETTO_Preprocess}
\usage{
AMARETTO_Preprocess(CancerSite, DataSetDirectories)
}
\arguments{
\item{DataSetDirectories}{}
}
\description{
Wrapper code that analyzes process TCGA GISTIC (CNV) and gene expression (rna-seq or microarray) data  via one call
}
\examples{
data('BatchData')
TargetDirectory <- file.path(getwd(),"Downloads/") # path to data download directory
CancerSite <- "LIHC"
DataSetDirectories <- AMARETTO_Download(CancerSite,TargetDirectory)
ProcessedData <- AMARETTO_Preprocess(CancerSite,DataSetDirectories)

}
