% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_functions.R
\name{AMARETTO_Run}
\alias{AMARETTO_Run}
\title{AMARETTO_Run}
\usage{
AMARETTO_Run(AMARETTOinit)
}
\arguments{
\item{AMARETTOinit}{List output from AMARETTO_Initialize().}
}
\value{
result
}
\description{
Function to run AMARETTO, a statistical algorithm to identify cancer drivers by integrating a variety of omics data from cancer and normal tissue.
}
\examples{
data('ProcessedDataLIHC')
AMARETTOinit <- AMARETTO_Initialize(MA_matrix = ProcessedDataLIHC$MA_matrix,
                                    CNV_matrix = ProcessedDataLIHC$CNV_matrix,
                                    MET_matrix = ProcessedDataLIHC$MET_matrix,
                                    NrModules = 5, VarPercentage = 50)

AMARETTOresults <- AMARETTO_Run(AMARETTOinit)
}
