% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_vizualize.R
\name{AMARETTO_VisualizeModule}
\alias{AMARETTO_VisualizeModule}
\title{AMARETTO_VisualizeModule
Function to visualize the gene modules}
\usage{
AMARETTO_VisualizeModule(AMARETTOinit, AMARETTOresults,
  CNV_matrix = NULL, MET_matrix = NULL, ModuleNr,
  SAMPLE_annotation = NULL, ID = NULL, order_samples = NULL)
}
\arguments{
\item{AMARETTOinit}{List output from AMARETTO_Initialize().}

\item{AMARETTOresults}{List output from AMARETTO_Run().}

\item{CNV_matrix}{CNV matrix, with genes in rows and samples in columns.}

\item{MET_matrix}{Methylation matrix, with genes in rows and samples in columns.}

\item{ModuleNr}{Module number to visualize}

\item{SAMPLE_annotation}{Matrix or Dataframe with sample annotation}

\item{ID}{Column used as sample name}

\item{order_samples}{Order samples in heatmap by mean or by clustering}
}
\description{
AMARETTO_VisualizeModule
Function to visualize the gene modules
}
