% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_vizualize.R
\name{AMARETTO_VisualizeModule}
\alias{AMARETTO_VisualizeModule}
\title{AMARETTO_VisualizeModule}
\usage{
AMARETTO_VisualizeModule(AMARETTOinit, AMARETTOresults,
  CNV_matrix = NULL, MET_matrix = NULL, ModuleNr,
  show_row_names = FALSE, SAMPLE_annotation = NULL, ID = NULL,
  order_samples = NULL)
}
\arguments{
\item{AMARETTOinit}{List output from AMARETTO_Initialize().}

\item{AMARETTOresults}{List output from AMARETTO_Run().}

\item{CNV_matrix}{CNV matrix, with genes in rows and samples in columns.}

\item{MET_matrix}{Methylation matrix, with genes in rows and samples in columns.}

\item{ModuleNr}{Module number to visualize.}

\item{show_row_names}{if TRUE, displays rownames of the gene expression matrix in the modules heatmap}

\item{SAMPLE_annotation}{Matrix or Dataframe with sample annotation.}

\item{ID}{Column used as sample name.}

\item{order_samples}{Order samples in heatmap by mean or by clustering.}
}
\description{
Function to visualize the gene modules
}
\examples{
data("ProcessedDataLIHC")
AMARETTOinit <- AMARETTO_Initialize(MA_matrix = ProcessedDataLIHC$MA_matrix,
                                    CNV_matrix = ProcessedDataLIHC$CNV_matrix,
                                    MET_matrix = ProcessedDataLIHC$MET_matrix,
                                    NrModules = 20, VarPercentage = 50)

AMARETTOresults <- AMARETTO_Run(AMARETTOinit)

ModuleNr <- 1 #define the module number to visualize

AMARETTO_VisualizeModule(AMARETTOinit = AMARETTOinit,AMARETTOresults = AMARETTOresults,
                         CNV_matrix = ProcessedDataLIHC$CNV,
                         MET_matrix = ProcessedDataLIHC$MET,
                         ModuleNr = ModuleNr)
}
