% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_vizualize.R
\name{AMARETTO_VisualizeModule}
\alias{AMARETTO_VisualizeModule}
\title{AMARETTO_VisualizeModule}
\usage{
AMARETTO_VisualizeModule(AMARETTOinit, AMARETTOresults,
  CNV_matrix = NULL, MET_matrix = NULL, ModuleNr,
  SAMPLE_annotation = NULL, ID = NULL, order_samples = NULL,
  printHM = FALSE)
}
\arguments{
\item{AMARETTOinit}{List output from AMARETTO_Initialize().}

\item{AMARETTOresults}{List output from AMARETTO_Run().}

\item{CNV_matrix}{CNV matrix, with genes in rows and samples in columns.}

\item{MET_matrix}{Methylation matrix, with genes in rows and samples in columns.}

\item{ModuleNr}{Module number to visualize.}

\item{SAMPLE_annotation}{Matrix or Dataframe with sample annotation.}

\item{ID}{Column used as sample name.}

\item{order_samples}{Order samples in heatmap by mean or by clustering.}

\item{printHM}{Boolean print heatmap directly}
}
\value{
result
}
\description{
Function to visualize the gene modules
}
\examples{
data('ProcessedDataLIHC')
AMARETTOinit <- AMARETTO_Initialize(MA_matrix = ProcessedDataLIHC$MA_matrix,
                                    CNV_matrix = ProcessedDataLIHC$CNV_matrix,
                                    MET_matrix = ProcessedDataLIHC$MET_matrix,
                                    NrModules = 5, VarPercentage = 50)

AMARETTOresults <- AMARETTO_Run(AMARETTOinit)

AMARETTO_VisualizeModule(AMARETTOinit = AMARETTOinit,AMARETTOresults = AMARETTOresults,
                         CNV_matrix = ProcessedDataLIHC$CNV,
                         MET_matrix = ProcessedDataLIHC$MET,
                         ModuleNr = 1)
}
