% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{ComBat_NoFiles}
\alias{ComBat_NoFiles}
\title{ComBat_NoFiles}
\usage{
ComBat_NoFiles(
  dat,
  saminfo,
  type = "txt",
  write = FALSE,
  covariates = "all",
  par.prior = TRUE,
  filter = FALSE,
  skip = 0,
  prior.plots = FALSE
)
}
\value{
result
}
\description{
ComBat_NoFiles
}
\keyword{internal}
