% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_initialize.R
\name{CreateRegulatorData}
\alias{CreateRegulatorData}
\title{CreateRegulatorData}
\usage{
CreateRegulatorData(MA_matrix = MA_matrix, CNV_matrix = NULL,
  MET_matrix = NULL, Driver_list = NULL, PvalueThreshold = 0.001,
  RsquareThreshold = 0.1, method = "union")
}
\value{
result
}
\description{
Determine potential regulator genes.
}
\keyword{internal}
