% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_initialize.R
\name{geneFiltering}
\alias{geneFiltering}
\title{geneFiltering}
\usage{
geneFiltering(Type, MAdata, Percentage)
}
\value{
result
}
\description{
Function to filter gene expression matrix
}
\keyword{internal}
