% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R,
%   R/amaretto_htmlreport_imaging.R
\name{GmtFromModules}
\alias{GmtFromModules}
\title{GmtFromModules}
\usage{
GmtFromModules(AMARETTOresults, driverGSEA)

GmtFromModules(AMARETTOresults, driverGSEA)
}
\arguments{
\item{AMARETTOresults}{List output from AMARETTO_Run().}

\item{driverGSEA}{if TRUE , module driver genes will also be added to module target genes for GSEA.}

\item{driverGSEA}{if TRUE , module driver genes will also be added to module target genes for GSEA.}

\item{AMARETTOresults}{List output from AMARETTO_Run().}
}
\value{
result

result
}
\description{
GmtFromModules

GmtFromModules
}
\keyword{internal}
