% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{GmtFromModules}
\alias{GmtFromModules}
\title{GmtFromModules}
\usage{
GmtFromModules(AMARETTOinit, AMARETTOresults, driverGSEA)
}
\value{
result
}
\description{
GmtFromModules
}
\keyword{internal}
