% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{GmtFromModules}
\alias{GmtFromModules}
\title{GmtFromModules}
\usage{
GmtFromModules(AMARETTOinit, AMARETTOresults)
}
\arguments{
\item{AMARETTOinit}{List output from AMARETTO_Initialize().}

\item{AMARETTOresults}{List output from AMARETTO_Run().}
}
\description{
Output a GMT file for the modules, read gmt, read description from gmt
}
\keyword{GmtFromModules(AMARETTOinit,AMARETTOresults)}
\keyword{internal}
