% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_html.R
\name{HyperGTestGeneEnrichment}
\alias{HyperGTestGeneEnrichment}
\title{Title}
\usage{
HyperGTestGeneEnrichment(gmtfile, testgmtfile, outfile, n_cluster,
  all_gene_address, show.overlapping.genes = FALSE,
  filter.genes = TRUE, show.unrecognized = FALSE)
}
\arguments{
\item{gmtfile}{}

\item{testgmtfile}{}

\item{outfile}{}

\item{n_cluster}{}

\item{all_gene_address}{}

\item{show.overlapping.genes}{}

\item{filter.genes}{}

\item{show.unrecognized}{}
}
\description{
Title
}
