% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{HyperGTestGeneEnrichment}
\alias{HyperGTestGeneEnrichment}
\title{Hyper Geometric Geneset Enrichement Test}
\usage{
HyperGTestGeneEnrichment(gmtfile, testgmtfile, NrCores,
  ref.numb.genes = 45956)
}
\arguments{
\item{gmtfile}{The gmt file with reference gene set.}

\item{testgmtfile}{The gmt file with gene sets to test. In our case, the gmt file of the modules.}

\item{NrCores}{Number of cores used for parallelization.}

\item{ref.numb.genes}{The total number of genes teste, standard equal to 45 956 (MSIGDB standard).}
}
\description{
Calculates the p-values for unranked gene set enrichment based on two gmt files as input and the hyper geometric test.
}
\keyword{internal}
