% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{HyperGTestGeneEnrichment}
\alias{HyperGTestGeneEnrichment}
\title{Hyper Geometric Geneset Enrichement Test}
\usage{
HyperGTestGeneEnrichment(gmtfile, testgmtfile, NrCores,
  ref.numb.genes = 45956)
}
\value{
result
}
\description{
Calculates the p-values for unranked gene set enrichment based on two gmt files as input and the hyper geometric test.
}
\keyword{internal}
