% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MethylStates}
\alias{MethylStates}
\title{MethylStates}
\format{An object of class \code{list} of length 25.}
\source{
\url{AMARETTO}
}
\usage{
MethylStates
}
\description{
A list of dataframes of Methyl States for example dataset from TCGA-LIHC.
}
\keyword{datasets}
