% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MsigdbMapping}
\alias{MsigdbMapping}
\title{MsigdbMapping}
\format{List}
\source{
\url{AMARETTO}
}
\usage{
MsigdbMapping
}
\description{
A dataset containing all MSIGDB pathways and their descriptions. .
}
\keyword{datasets}
