% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R
\name{plot_run_history}
\alias{plot_run_history}
\title{Title plot_run_history}
\usage{
plot_run_history(AMARETTOinit, AMARETTOResults)
}
\arguments{
\item{AMARETTOinit}{}

\item{AMARETTOResults}{}
}
\value{
plot
}
\description{
Title plot_run_history
}
\examples{
 plot_run_history(AMARETTOinit,AMARETTOResults)
}
\keyword{internal}
