% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{Preprocess_MAdata}
\alias{Preprocess_MAdata}
\title{Preprocess_MAdata}
\usage{
Preprocess_MAdata(CancerSite = CancerSite, MAEO_ge = MAEO_ge,
  BatchData = BatchData)
}
\value{
result
}
\value{
result
}
\description{
Preprocess_MAdata
}
\keyword{internal}
