% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProcessedDataLAML}
\alias{ProcessedDataLAML}
\title{ProcessedDataLAML}
\format{An object of class \code{list} of length 3.}
\source{
\url{AMARETTO}
}
\usage{
ProcessedDataLAML
}
\description{
A list of dataframes of processed toy example dataset from TCGA-LAML.
}
\keyword{datasets}
