% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_initialize.R
\name{read_gct}
\alias{read_gct}
\title{read_gct}
\usage{
read_gct(file_address)
}
\arguments{
\item{file_address}{}
}
\value{
result
}
\description{
Function to turn a .gct data files into a matrix format
}
\examples{
data_matrix<-read_gct(my_gctFile.gct)
}
