% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{Save_CancerSite}
\alias{Save_CancerSite}
\title{Save_CancerSite}
\usage{
Save_CancerSite(CancerSite, TargetDirectory, DataSetDirectories,
  ProcessedData)
}
\arguments{
\item{ProcessedData}{}
}
\value{
result
}
\description{
Save_CancerSite
}
\keyword{internal}
