% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_BatchCorrection_MolecularData}
\alias{TCGA_BatchCorrection_MolecularData}
\title{TCGA_BatchCorrection_MolecularData}
\usage{
TCGA_BatchCorrection_MolecularData(
  GEN_Data = GEN_Data,
  BatchData = BatchData,
  MinInBatch = MinInBatch
)
}
\value{
result
}
\description{
TCGA_BatchCorrection_MolecularData
}
\keyword{internal}
