% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_GENERIC_CheckBatchEffect}
\alias{TCGA_GENERIC_CheckBatchEffect}
\title{TCGA_GENERIC_CheckBatchEffect}
\usage{
TCGA_GENERIC_CheckBatchEffect(GEN_Data, BatchData)
}
\value{
result
}
\value{
result
}
\description{
TCGA_GENERIC_CheckBatchEffect
}
\keyword{internal}
