% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_GENERIC_CleanUpSampleNames}
\alias{TCGA_GENERIC_CleanUpSampleNames}
\title{TCGA_GENERIC_CleanUpSampleNames}
\usage{
TCGA_GENERIC_CleanUpSampleNames(GEN_Data, IDlength = 12)
}
\arguments{
\item{IDlength}{}
}
\description{
TCGA_GENERIC_CleanUpSampleNames
}
\keyword{internal}
