% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_preprocess.R
\name{TCGA_GENERIC_MergeData}
\alias{TCGA_GENERIC_MergeData}
\title{TCGA_GENERIC_MergeData}
\usage{
TCGA_GENERIC_MergeData(NewIDListUnique, DataMatrix, MergeMethod)
}
\value{
result
}
\description{
TCGA_GENERIC_MergeData
}
\keyword{internal}
